Transfer project ownership (#1600)
Project owners and site admins can now transfer ownership of a project to another user. Site admins can now delete projects as well as project owners.

+++ src/angular-app/languageforge/lexicon/shared/share-with-others/user-management.component.ts
@@ -1,7 +1,8 @@
 import * as angular from 'angular';
 import { ProjectService } from '../../../../bellows/core/api/project.service';
 import { UserService } from '../../../../bellows/core/api/user.service';
-import { Session } from '../../../../bellows/core/session.service';
+import { Session, SessionService } from '../../../../bellows/core/session.service';
+import { ModalService } from '../../../../bellows/core/modal/modal.service';
 import { UtilityService } from '../../../../bellows/core/utility.service';
 import { Project, ProjectRole } from '../../../../bellows/shared/model/project.model';
 import { User } from '../../../../bellows/shared/model/user.model';
@@ -19,13 +20,16 @@ export class UserManagementController implements angular.IController {
   anonymousUserRoles: ProjectRole[];
   memberRoles: ProjectRole[];
 
-  static $inject = ['$q', 'projectService', 'userService'];
+
+  static $inject = ['$q', 'projectService', 'userService', 'sessionService','modalService'];
   constructor(
     private readonly $q: angular.IQService,
     private readonly projectService: ProjectService,
-    private readonly userService: UserService) { }
+    private readonly userService: UserService,
+    private readonly sessionService: SessionService,
+    private readonly modal: ModalService) { }
 
-  $onInit(): void {
+  async $onInit(): Promise<void> {
     // TODO: actually hook anonymousUserRole up to the backend
     this.project.anonymousUserRole = LexRoles.NONE.key;
 
@@ -42,13 +46,17 @@ export class UserManagementController implements angular.IController {
       LexRoles.OBSERVER_WITH_COMMENT,
       LexRoles.OBSERVER
     ];
+
+    await this.sessionService.getSession().then((s) => {
+        this.session = s;
+      });
   }
 
-  userIsCurrentUser(user: User) {
+  userIsCurrentUser(user: User): boolean {
     return this.session.data.username === user.username;
   }
 
-  userIsOwner(user: User) {
+  userIsOwner(user: User): boolean {
     return user.id === this.project.ownerRef.id;
   }
 
@@ -56,6 +64,24 @@ export class UserManagementController implements angular.IController {
     this.projectService.updateUserRole($event.target.id, $event.roleDetail.role.key).then(() => {
       this.loadMemberData();
     });
+  }
+
+  onOwnershipTransfer($event: {target: Partial<User>}) {
+    const ownershipTransferMessage = 'Are you sure you want to transfer ownership of <b>' + this.project.projectName + '</b> to <b>' + $event.target.username + '</b>?';
+    this.modal.showModalSimple('Transfer project ownership', ownershipTransferMessage, 'Cancel', 'Transfer ownership to ' + $event.target.username).then(() => {
+      var newOwnerId = $event.target.id;
+      this.projectService.transferOwnership(newOwnerId).then(() => {
+        this.loadMemberData();
+        this.project.ownerRef.id = newOwnerId;
+      });
+    })
+  }
+
+  currentUserIsOwnerOrAdmin(): boolean {
+    if (typeof this.project.ownerRef == 'object') {
+      return (this.project.ownerRef.id === this.session.data.userId) || this.sessionService.userIsAdmin();
+    }
+    return false;
   }
 
   onSpecialRoleChanged($event: {roleDetail: RoleDetail, target: string}) {
@@ -84,8 +110,11 @@ export class UserManagementController implements angular.IController {
   }
 
   removeUser(user: User) {
-    this.projectService.removeUsers([user.id]).then(() => {
-      this.loadMemberData();
+    const removeUserMessage = 'Are you sure you want to remove <b>' + user.username + '</b> from the <b>' + this.project.projectName + '</b> project?';
+    this.modal.showModalSimple('Remove user', removeUserMessage, 'Cancel', 'Remove ' + user.username).then(() => {
+      this.projectService.removeUsers([user.id]).then(() => {
+        this.loadMemberData();
+      });
     });
   }
